Move the Network.MessageQueue class into the Node class
Since each MessageQueue actually belongs to a Node (each Node has a MessageQueue). The inner class and the field are declared private so that the application code cannot use it. Two package-private methods are available to allow the Network class (but not the application code) to interact with the MessageQueue.
