+++ src/main/java/dev/oxoo2a/sim4da/Network.java
@@ -1,27 +1,17 @@
 package dev.oxoo2a.sim4da;
 
-import java.util.LinkedList;
-import java.util.Random;
-import java.util.concurrent.Semaphore;
-
 import dev.oxoo2a.sim4da.Message.MessageType;
 
 public class Network {
     
-    private static final Random RANDOM = new Random();
-    
     private final int numberOfNodes;
     private final Tracer tracer;
-    private final MessageQueue[] messageQueues;
+    private final Node[] nodes;
     
-    //private static Logger logger = Logger.getRootLogger();
-    
-    public Network(int numberOfNodes, Tracer tracer) {
+    public Network(int numberOfNodes, Node[] nodes, Tracer tracer) {
         this.numberOfNodes = numberOfNodes;
+        this.nodes = nodes;
         this.tracer = tracer;
-        messageQueues = new MessageQueue[numberOfNodes];
-        for (int i = 0; i<numberOfNodes; i++)
-            messageQueues[i] = new MessageQueue();
     }
     
     public int getNumberOfNodes() {
@@ -39,7 +29,7 @@ public class Network {
         }
         tracer.emit("Unicast:%d->%d", senderId, receiverId);
         Message raw = new Message(senderId, receiverId, MessageType.UNICAST, message);
-        messageQueues[receiverId].put(raw);
+        nodes[receiverId].putInMessageQueue(raw);
     }
     
     public void broadcast(int senderId, String message) {
@@ -52,7 +42,7 @@ public class Network {
         for (int l = 0; l<numberOfNodes; l++) {
             if (l==senderId) continue;
             raw.receiverId = l;
-            messageQueues[l].put(raw);
+            nodes[l].putInMessageQueue(raw);
         }
     }
     
@@ -61,49 +51,11 @@ public class Network {
             System.err.printf("Network::receive: unknown receiver id %d\n", receiverId);
             return null;
         }
-        Message m = messageQueues[receiverId].await();
+        Message m = nodes[receiverId].awaitFromMessageQueue();
         if (m!=null) {
             String m_type = m.type==MessageType.BROADCAST ? "Broadcast" : "Unicast";
             tracer.emit("Receive %s:%d<-%d", m_type, m.receiverId, m.senderId);
         }
         return m;
-    }
-    
-    public void stop() {
-        for (MessageQueue queue : messageQueues)
-            queue.stop();
-    }
-    
-    private static class MessageQueue {
-        private final LinkedList<Message> queue = new LinkedList<>();
-        private final Semaphore awaitMessage = new Semaphore(0);
-        private boolean stop = false;
-        public void put(Message message) {
-            synchronized (queue) {
-                queue.addLast(message);
-                awaitMessage.release();
-            }
-        }
-        public Message await() {
-            while (!stop) {
-                try {
-                    awaitMessage.acquire();
-                    synchronized (queue) {
-                        if (!queue.isEmpty()) {
-                            // Return a random message in queue avoiding FIFO order
-                            if (queue.size() == 1)
-                                return queue.removeFirst();
-                            int c = RANDOM.nextInt(queue.size());
-                            return queue.remove(c);
-                        }
-                    }
-                } catch (InterruptedException ignored) {}
-            }
-            return null; // Simulation time ended before a message was received
-        }
-        public void stop() {
-            stop = true;
-            awaitMessage.release();
-        }
     }
 }
