+++ src/main/java/dev/oxoo2a/sim4da/Node.java
@@ -1,8 +1,15 @@
 package dev.oxoo2a.sim4da;
 
+import java.util.LinkedList;
+import java.util.Random;
+import java.util.concurrent.Semaphore;
+
 public abstract class Node {
     
+    private static final Random RANDOM = new Random();
+    
     protected final int id;
+    private final MessageQueue messageQueue = new MessageQueue();
     private Network network;
     private Tracer tracer;
     private final Thread thread = new Thread(this::run);
@@ -26,6 +33,7 @@ public abstract class Node {
     
     public void stop () {
         stop = true;
+        messageQueue.stop(); // Stop waiting in receive
         try {
             thread.join();
         } catch (InterruptedException ignored) {}
@@ -59,6 +67,49 @@ public abstract class Node {
         return network.receive(id);
     }
     
+    // The two methods below are package-private because they shouldn't be used by application code
+    
+    void putInMessageQueue(Message message) {
+        messageQueue.put(message);
+    }
+    
+    Message awaitFromMessageQueue() {
+        return messageQueue.await();
+    }
+    
     // Module implements basic node functionality
     protected abstract void run();
+    
+    private static class MessageQueue {
+        private final LinkedList<Message> queue = new LinkedList<>();
+        private final Semaphore awaitMessage = new Semaphore(0);
+        private boolean stop = false;
+        private void put(Message message) {
+            synchronized (queue) {
+                queue.addLast(message);
+                awaitMessage.release();
+            }
+        }
+        private Message await() {
+            while (!stop) {
+                try {
+                    awaitMessage.acquire();
+                    synchronized (queue) {
+                        if (!queue.isEmpty()) {
+                            // Return a random message in queue avoiding FIFO order
+                            if (queue.size() == 1)
+                                return queue.removeFirst();
+                            int c = RANDOM.nextInt(queue.size());
+                            return queue.remove(c);
+                        }
+                    }
+                } catch (InterruptedException ignored) {}
+            }
+            return null; // Simulation time ended before a message was received
+        }
+        public void stop() {
+            stop = true;
+            awaitMessage.release();
+        }
+    }
 }
